refactor: move Render photo function to higher level component to call dispatch just once per search photos request
