+++ react-components/src/components/pages/MainPage/components/SearchResult.tsx
@@ -1,10 +1,13 @@
 import React from 'react';
 import { IPhoto, IPhotos, IResponse, RootState } from 'types';
 import isEmpty from 'helpers/isEmpty';
-import RenderPhoto from './RenderPhoto';
-import { useSelector } from 'react-redux';
+import { useDispatch, useSelector } from 'react-redux';
+import PhotoCard from './PhotoCard';
+import getDetailContent from '../functions/getDeteilContent';
+import { saveDetailContent } from 'store/slices/detailSlice';
 
 function SearchResult(): JSX.Element {
+  const dispatch = useDispatch();
   const responseObj = useSelector((state: RootState): IResponse | {} => state.photos.response);
   const renderSearchResult = (): JSX.Element[] => {
     let photosArr: IPhoto[] = [];
@@ -19,8 +22,24 @@ function SearchResult(): JSX.Element {
 
     let photoCardsArr: Array<JSX.Element> = [];
     if (photosArr.length > 0) {
-      photoCardsArr = photosArr.map((photo: IPhoto | []): JSX.Element => {
-        return RenderPhoto(photo as IPhoto);
+      photoCardsArr = photosArr.map((photo: IPhoto): JSX.Element => {
+
+        const srcMedium = `https://live.staticflickr.com/${photo.server}/${photo.id}_${photo.secret}_z.jpg`;
+        const srcLarge = `https://live.staticflickr.com/${photo.server}/${photo.id}_${photo.secret}_c.jpg`;
+
+        return (
+          <PhotoCard
+            onClick={(event) => {
+              event.stopPropagation();
+              const detailContent = getDetailContent(photo, srcLarge);
+              dispatch(saveDetailContent({ newDetailState: detailContent }));
+            }}
+            key={+photo.id}
+            src={srcMedium}
+            title={photo.title}
+            owner={photo.owner}
+          />
+        );
       });
     }
